+++ packages/autocomplete-shared/src/index.ts
@@ -1,5 +1,7 @@
 export * from './createRef';
 export * from './debounce';
+export * from './generateAutocompleteId';
+export * from './getAttributeValueByPath';
 export * from './getItemsCount';
 export * from './invariant';
 export * from './isEqual';